requirements.txt
django-post-office==3.4.1
django==2.2.10
djangorestframework==3.11.0
factory-boy==2.12.0
faker==4.1.1
htmlentities==0.3.0
mysqlclient==1.4.2.post1
django-extensions==2.2.8
